Update Portal to authenticate using completely external IdP
# Previously the Federated IdP was selected from the Portal and it was sent to the IdP. In the new approach, the FIdP selection itself had been moved to the IdP
